tests/wpt/tests/html/infrastructure/urls/terminology-0/multiple-base.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>document base URL: multiple base elements</title>
<base target="_blank" >
<base href="http://{{domains[www]}}:{{ports[http][0]}}/">
<base href="http://{{domains[www1]}}:{{ports[http][0]}}/">
<base href="http://{{domains[www2]}}:{{ports[http][0]}}/">
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  test(function(){
    var base = document.querySelectorAll("base");
    assert_equals(document.baseURI, document.querySelectorAll("base[href]")[0].href);
  }, "If there are multiple <base> elements, the document base URL is the frozen base URL of the first one that has an href attribute");
</script>
